requirements.txt
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.1
FormEncode==2.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
m-lib==3.1.1
m-lib.defenc==1.1.0.post2
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
PyDispatcher==2.0.7
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.2
six==1.17.0
sqlconvert==0.3.0
SQLObject==3.13.0
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
